lionlist/requirements.txt
Django==1.4.1
South==0.8.2
Unipath==1.0
argparse==1.2.1
boto==2.14.0
cov-core==1.7
coverage==3.6
distribute==0.6.24
django-autoslug==1.7.1
django-braces==1.0.0
django-dajax==0.9.2
django-dajaxice==0.5.5
django-haystack==2.0.0
django-password-reset==0.5
django-registration==1.0
django-ses==0.4.1
django-smart-selects==1.0.9
django-widget-tweaks==1.3
mock==1.0.1
py==1.4.15
pyelasticsearch==0.6
pytest==2.3.5
pytest-cov==1.6
pytest-django==2.3.0
requests==1.2.3
simplejson==3.3.0
six==1.3.0
wsgiref==0.1.2
